Util functions moved in util.rs
