+++ src/lib/util.rs
@@ -3,7 +3,8 @@ use crate::lib::term::{
     writegreen, 
     clear
 };
- 
+use rand::prelude::*;
+use std::io; 
 /*
 All utils(exit process,...)    
 */
@@ -26,4 +27,21 @@ pub fn intize(input: String) -> i32 {
           std::process::exit(1); 
        }
     } 
- }
+}
+pub fn genrandom() -> i16 {
+    return rand::thread_rng().gen_range(1, 50); 
+}
+ 
+pub fn getuserinput() -> String {
+    let mut input = String::new(); 
+    match io::stdin().read_line(&mut input) {
+        Ok(_n) => {
+          return input; 
+        }
+        Err(e) => {
+          println!("Erreur lors de la récupération de l'entrée! ({:?})", e ); 
+          std::process::exit(1);
+       }
+    }
+}
+
